pyproject.toml
[project]
name = "weavops"
version = "0.0.2"
description = "WeavOps bridges GitOps with MCP-powered AI agents, automatically weaving your Git source of truth into autonomous, context-aware operations"
readme = "README.md"
authors = [{ name = "Yan-Ke Guo", email = "hi@gyk.me" }]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: System :: Systems Administration",
]

[project.urls]
Homepage = "https://github.com/yankeguo/weavops"
Repository = "https://github.com/yankeguo/weavops"
Issues = "https://github.com/yankeguo/weavops/issues"

[project.scripts]
weavops = "weavops:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"
